note update api modified and delete api done

+++ notes/serializers.py
@@ -23,12 +23,12 @@ class NoteSerializer(ModelSerializer):
         fields = ['text', 'note_tag']
 
 
-class NoteCreateUpdateSerializer(ModelSerializer):
+class NoteCreateUpdateDeleteSerializer(ModelSerializer):
     tags = serializers.ListField(child=serializers.CharField(), write_only=True)
 
     class Meta:
         model = Note
-        fields = ['text', 'tags']
+        fields = ['text', 'tags', 'is_archived']
 
     def create(self, validated_data):
         tags = validated_data.pop('tags')
@@ -45,15 +45,18 @@ class NoteCreateUpdateSerializer(ModelSerializer):
         return note_instance
 
     def update(self, instance, validated_data):
-        NoteTag.objects.filter(created_by=instance.created_by, note=instance ).delete()
-        tags = validated_data.pop('tags')
-        if tags:
-            for tag in tags:
-                tag_obj_exist = Tag.objects.filter(name=tag).exists()
-                if tag_obj_exist:
-                    NoteTag.objects.create(tag=Tag.objects.get(name=tag), note=instance, created_by=self.context['request'].user)
-                if not tag_obj_exist:
-                    tag_obj = Tag.objects.create(name=tag)
-                    NoteTag.objects.create(tag=tag_obj, note=instance,
-                                           created_by=self.context['request'].user)
+
+        if validated_data.get('is_archived') == False:
+            NoteTag.objects.filter(created_by=instance.created_by, note=instance ).delete()
+            tags = validated_data.pop('tags')
+            if tags:
+                for tag in tags:
+                    tag_obj_exist = Tag.objects.filter(name=tag).exists()
+                    if tag_obj_exist:
+                        NoteTag.objects.create(tag=Tag.objects.get(name=tag), note=instance, created_by=self.context['request'].user)
+                    if not tag_obj_exist:
+                        tag_obj = Tag.objects.create(name=tag)
+                        NoteTag.objects.create(tag=tag_obj, note=instance,
+                                               created_by=self.context['request'].user)
+            validated_data.update({"created_by": self.context['request'].user})
         return super().update(instance, validated_data)
